# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Medical App API
  version: 1.0.0
  description: REST API for the Medical App application
servers:
  - url: /api
    description: Default server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HttpOnly access token cookie set on login
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid credentials
    MessageResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
              example: Success
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh the access token
      operationId: refreshToken
      description: Reads the `refreshToken` cookie and issues new auth cookies.
      responses:
        "200":
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Missing or invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...